+++ src/components/Agreement/Agreement.js
@@ -1,4 +1,4 @@
-import React from 'react'
+import React, { useState } from 'react'
 import { Button, Header } from '@aragon/ui'
 import LayoutGutter from '../Layout/LayoutGutter'
 import LayoutLimiter from '../Layout/LayoutLimiter'
@@ -9,23 +9,43 @@ import AgreementDocument from './AgreementDocument'
 import LayoutBox from '../Layout/LayoutBox'
 import LayoutColumns from '../Layout/LayoutColumns'
 import LoadingSection from '../Loading/LoadingSection'
+import MultiModal from '../MultiModal/MultiModal'
+import SignAgreementModal from '../ModalFlows/SignAgreementModal/SignAgreementModal'
 import { useAgreement } from '../../providers/Agreement'
+import { useWallet } from '../../providers/Wallet'
 
 const Agreement = React.memo(function Agreement() {
+  const { account } = useWallet()
+  const [signAgreementVisible, setSignAgreementVisible] = useState(false)
   const { agreementDetails, loading } = useAgreement()
 
   return (
-    <LayoutGutter>
-      <LayoutLimiter>
-        <Header
-          primary="Agreement"
-          secondary={<Button mode="strong" label="Sign Agreement" disabled />}
-        />
-        <LoadingSection title="Loading agreement" loading={loading}>
-          <AgreementLayout agreementDetails={agreementDetails} />
-        </LoadingSection>
-      </LayoutLimiter>
-    </LayoutGutter>
+    <>
+      <LayoutGutter>
+        <LayoutLimiter>
+          <Header
+            primary="Agreement"
+            secondary={
+              <Button
+                mode="strong"
+                label="Sign Agreement"
+                disabled={!account}
+                onClick={() => setSignAgreementVisible(true)}
+              />
+            }
+          />
+          <LoadingSection title="Loading agreement" loading={loading}>
+            <AgreementLayout agreementDetails={agreementDetails} />
+          </LoadingSection>
+        </LayoutLimiter>
+      </LayoutGutter>
+      <MultiModal
+        visible={signAgreementVisible}
+        onClose={() => setSignAgreementVisible(false)}
+      >
+        <SignAgreementModal />
+      </MultiModal>
+    </>
   )
 })
