+++ src/components/ModalFlows/ModalButton.js
@@ -1,0 +1,11 @@
+import styled from 'styled-components'
+import { Button, GU } from '@aragon/ui'
+
+const ModalButton = styled(Button).attrs({
+  mode: 'strong',
+  wide: true,
+})`
+  margin-top: ${2 * GU}px;
+`
+
+export default ModalButton
